# optimizing mimetype serialization