kj::strTree results in plenty of heap allocation. Using capacity-bound string buffer to do only one allocation per serialize

before:

```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
Mimetype::ParseAndSerialize       2.18 us         2.18 us       324082
Mimetype::Serialize              0.351 us        0.351 us      1982953

```

after:

```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
Mimetype::ParseAndSerialize       1.48 us         1.48 us       456517
Mimetype::Serialize              0.114 us        0.114 us      6175365
```