@mikea 2023-08-24
updated string buffer to be unbounded and to grow on the heap. ptal @kentonv @jasnell