@irvinebroque 2023-08-24
Does new benchmarking tooling we added gives us a way to fail CI if a given benchmark regresses more than a given percentage? (like if something shot up 400%)