@mikea 2023-08-24
> Does new benchmarking tooling we added gives us a way to fail CI if a given benchmark regresses more than a given percentage? (like if something shot up 400%)

eventually this will be possible. Right now it is just an instrument to measure a single change impact. I plan to start collecting these on TC for a longer term trend